backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.2.1

# PDF processing
pymupdf>=1.27.0
pdfplumber==0.11.0

# AI / embeddings
sentence-transformers==2.7.0
google-generativeai==0.5.4

# HTTP
httpx>=0.28.1 

# Vector DB
endee==0.1.25

# Env
python-dotenv==1.0.1